cmd-report: switch to offline mode if ask-configs are ignored
When a user generates a system report, the configuration may indicate
that username/password options are in ask mode (which in turn are used
to validate connectivity to the Confluence instance). If a user opts out
of providing valid entries for these, the report process fails to handle
these lack-of-input cases.

In the event that a user does not provide values for an ask-enabled
configuration, automatically switch to offline mode.

Signed-off-by: James Knight <james.d.knight@live.com>