cmd-report: log full traceback of detected exceptions
When handling error cases in report generation (such as a syntax error
in a configuration or a bug in the extension), only a limited summary
message is printed. Instead, dump the full stack trace to provide a user
more details (which in turn, can be used to debug their setups or help
in a report for an issue).

Signed-off-by: James Knight <james.d.knight@live.com>