publisher: default can_labels detection
The `confluence_adv_restricted` configuration entry may not be
configured with a default value for a publisher's initialization call
(e.g. for builders which may not apply a default configuration hook).
Improve handling when attempting to cache a `can_labels` state by
ensuring `confluence_adv_restricted` is set before attempting to use it
for consideration.

Signed-off-by: James Knight <james.d.knight@live.com>