provide hints for users using a space name over a space key
Contrary to the name of the `confluence_space_name` configuration, this
option is designed to accept a space's key value instead of a space's
name value. If a user does specify a space's name in this entry, attempt
to provide a hint on the key value that should be configured. In the
event that space key is not found, attempt to search for spaces named
with the value configured in `confluence_space_name`. If a matching
space is detect, inform the user of the key value that should be used
instead.

This commit also tweaks the message to the user to be a bit more clear
that a space key value is desired.

Signed-off-by: James Knight <james.d.knight@live.com>