publisher: better reflect new attachments in dry-runs
If a user performs a dry-run test which includes new attachments for
pages that do not exist, the run will report several warning messages
for these attachments:

    WARNING: cannot publish asset since publishing point cannot be found (<file>): <document>

Since this is a dry-run, there is not a need to validate a publishing
point for an attachment. In the event that a publishing point is not
detected, handle it in a way that represents a new attachment.

Signed-off-by: James Knight <james.d.knight@live.com>