cmd-wipe: reflect dry-run mode
The wipe command respect the `confluence_publish_dryrun` option, if it
is set in the active configuration. This may not be obvious to users
invoking a dry-run wipe event where it will indicate that pages are
being removed, but they are not.

If the dry-run option is enabled, present this in the configuration
information given to a user. Also, better present "processed" pages by
avoiding period-character status updates in this mode.

Signed-off-by: James Knight <james.d.knight@live.com>