tox: explicitly pin docutils<0.18 for older sphinx versions
Older versions of Sphinx do not support the recent docutils v0.18
release. Explicitly configuring the test environment to enforce this for
the older versions being tested.

Note that in Sphinx v4+, the supported docutils version is listed in its
setup definition (i.e. does not need to be managed here).

Signed-off-by: James Knight <james.d.knight@live.com>