inject xml declaration if an svg attachment does not have one
Confluence will accept the upload of an SVG without an XML declaration;
however, it will render the macro as corrupted image. To properly render
SVG images on Confluence, the declaration is required. This commit
adjusts the publishing process to inject an XML declaration in an SVG
file before attaching it to a page.

Signed-off-by: James Knight <james.d.knight@live.com>