storage: improve image path detection when applying a scaled width
When using a `scale` option for an image, the image size will be
calculated to help calculate a proper width value to apply for an image.
The path for an image uses the builder's source directory as the root
for all images, which is not always the case. Adjust the logic to use
the full path determined by asset manager to ensure an existing image
file has a proper path to determine sizing data on.

Signed-off-by: James Knight <james.d.knight@live.com>