correct xml-declaration injection for svgs
A previous commit [1] added support to inject XML declarations for SVGs
which may be missing them. When the change was prepared, a switch in the
development version from a `<svg` check to an `<?xml` check did not
properly update the condition -- which now leads to injecting XML
declaration on SVGs which already have them; correcting this.

[1]: dcc7f027f00e0909851586c9c473e6c30fd7f114

Signed-off-by: James Knight <james.d.knight@live.com>