chunk a generated hash-comment for assets
When assets are published, the comment field is assigned a key-hash
value which can be queried for later to avoid re-uploading assets on
multiple publishing events. This generated value is a single long
string, which can cause display issues in the attachment area (text
overflowing specific asset actions). Attempt to minimize this by
chunking the hash value into parts.

Signed-off-by: James Knight <james.d.knight@live.com>