support genindex generation/processing
Provides support in the generation of an index for a documentation set.
Indexes will now be generated when either the user specifies
`confluence_use_index` or if a `genindex` document is referenced in the
toctree. Users can also opt out of any index generation by explicitly
configuring `confluence_use_index` to `False`.

Signed-off-by: James Knight <james.d.knight@live.com>