support search generation
Provides support in the generation of a search for a documentation set.
A search document will now be generated when either the user specifies
`confluence_include_search` or if a `search` document is referenced in
the toctree. Users can also opt out of any search document generation by
explicitly configuring `confluence_include_search` to `False`.

Signed-off-by: James Knight <james.d.knight@live.com>