skip processing for assets on a doctree-resolve
The following removes the hook to re-check for assets after a doctree
has been resolved. While it is still possible for new assets that need
to be processed after a doctree has been resolved, there can still be
events later on in the process where additional node changes can still
result in assets needing to be processed for again. Support for this is
handled in the translator class when a given asset-based node (i.e.
images or downloads) will trigger a node processing for asset on-demand.

Since there is no major difference to looking for standalone assets on a
doctree resolve event or when the translator needs to process a node,
dropping the event hook for simplicity in the implementation.

Signed-off-by: James Knight <james.d.knight@live.com>