tests: adjust svg doctype file to be more unique
Unit tests are failing under a Linux environment since when a XML
declaration correction is applied to `svg-doctype.svg`, it will result
in a matching hash to the `svg.svg` file. The unit test expects a
generated file for comparision, and fails since it checks the output
folder for the file.

Adjusting the `svg-doctype.svg` so that after being manipulated, it will
not be generated in a way which matches any other SVG file.

Signed-off-by: James Knight <james.d.knight@live.com>