translator: restore standalone fallback docname for assets
When processing the fallback condition for image/file nodes (for
standalone assets), a change was added [1] which removed passing the
active document into the asset manager (so it knows the proper document
to attach an asset to). The removal of these, while valid (or
irrelevant) for a `singleconfluence` builder is still needed for the
standard `confluence` builder; restoring.

[1]: 44b6cd19b60c5cb07218ad025e1ea8340a330d4a

Signed-off-by: James Knight <james.d.knight@live.com>