apply lengths for svgs with only a viewbox
SVGs which only have a `viewbox` attribute may not properly render on
Confluence (e.g. an SVG may visually load for a moment and the hide
itself). By specifying an explicit width/height value inside SVGs
provides a more consistent user experience -- adjusting SVG processing
so that when a viewbox-only SVG is detected, use these values as the
target width/height options to apply (which should set the attributes on
the final SVG image).

Signed-off-by: James Knight <james.d.knight@live.com>