doc: add upgrade flag for installation commands
Adjusting the installation commands to include the upgrade option. This
is to provide a common command for users installing or attempting to
upgrade -- with the intent to minimize for new users of pip who may not
know how to upgrade easily.

This also changes the setuptools calls to use the shorter upgrade flag.

Signed-off-by: James Knight <james.d.knight@live.com>