add a publishing pre-check for self-referenced ancestor request
Adding a check to help detect if a request has been made to publish a
page as a child of itself. This issue should only ever happen if a
configuration sets `confluence_parent_page` with the same value of
the title page for the documentation's `root_doc`. On Confluence Cloud,
a publish attempt for this case provides a descriptive message
"Cannot add an existing ancestor as a child!"; although on Confluence
self-hosted solutions, it can report a non-descriptive 500 error.

Adding this message should provide a more helpful description of the
issue, and also avoid invoking an invalid request to a Confluence
instance.

Signed-off-by: James Knight <james.d.knight@live.com>