improve support for handling percentage lengths on images
Percentage length values show some interesting results when used with
Confluence. Originally, this extension just allowed length values to
pass-through. Testing over time showed that using percentages would
sometimes cause the rendering of an image to throw a macro error. With
recent work related to SVG length handling improvements, percentages
were assumed to be just a not-supported unit type. However, additional
testing showed that percentage values did work for external images, and
only broke for images attached to a page.

This commit first restores the ability to use percentage lengths. They
should still be able to be set and applied for any external images. To
help deal with internal images with a percentage value, these images
will be wrapped with an inlined div element with the percentage widths
applied on the container (which appears to have a near-matching result
as if they were applied on the images themselves).

While the `ac:image` macro does indicate there is some hints to apply
class and style options, this only applied to the `img` tag itself.
Images on Confluence are also wrapped by a `span` tag which have a
`confluence-embedded-file-wrapper` class set. This class restricts the
width of an image (hence why the container has been used instead).

Signed-off-by: James Knight <james.d.knight@live.com>