translator-storage: remove undesired spacing on index pages
Index pages will generate horizontal bars for the letter links at the
top of the page. When support to add a page generation notice at the top
of these pages, the template used applied a margin offset which was
primarily for non-index pages. Adjusting the horizontal line in these
cases to not have the extra margin at the bottom.

Signed-off-by: James Knight <james.d.knight@live.com>