add ancestor check
Adding a pre-check to generate a detailed exception when a page is being
updated as a child of itself. When a root page is first upload,
ancestors of the page will be fetched and marked as "reserved". If a
future page update request attempts to move a root page's ancestor under
the root page's tree, the `ConfluencePublishAncestorError` exception
will be raised.

See also: e4368927c4ff509d5765ed3459a16c18d9dbe3a6

Signed-off-by: James Knight <james.d.knight@live.com>