improve node definitions
Adjusts a series of custom node types to inherit from more proper
definitions. When attempting to perform JIRA directive substitutions, it
was observed that the substitution declarations were not considered
valid. One of the main reasons for this is that substitutions only work
on inlined nodes, which neither JIRA node type inherited from.
Consulting the docutils documentation on elements [1], nodes have been
updated to inherit from more appropriate types.

While JIRA nodes are marked as inline now, another issue with their
definitions is that the internal `params` attribute would not be
properly copied when needed -- specifically, when substitutions are at
play. To improve support in tracking Confluence-specific parameter
key-value pairs (for macro parameter injection), this information will
now be stored inside a Element node's attributes dictionary. The
`attributes` attribute is properly managed/copied when needed, which
should ensure Confluence-specific parameters will be copied as well.
This will be managed inside a new `ConfluenceParams` node type, which
JIRA nodes (as well as the source-link node) can inherit from.

[1]: https://docutils.sourceforge.io/docs/ref/doctree.html

Signed-off-by: James Knight <james.d.knight@live.com>