storage-translator: remove information macro from figure structure
The original implementation added to support figures would wrap the
contents inside an informational macro. This was a styling preference at
the time (to produce a shaded background); however, it is not consistent
with how figures are generated with other builds (like HTML). Adjusting
the figure implementation to do what is common -- a simple paragraph
grouping on the figure structure.

Signed-off-by: James Knight <james.d.knight@live.com>