tests: improve the delay when building a mock confluence instance
It has been observed in some cases where attempting to mock a Confluence
instance, the testing implementation may fail to connect to the instance
with a connection reset event. The original implementation would attempt
to yield the context by sleeping with a zero count; however, this does
not ensure that the HTTP serving thread is ready set. To improve the
chance that the test HTTP server is ready, first wait on the thread
being started by synchronizing an `Event` type between the serving
thread and the testing context. After confirming the thread is running,
yield the context a bit to improve the chance of a `serve_forever`
invoke to be ready.

Signed-off-by: James Knight <james.d.knight@live.com>