publisher: forward explicit space type when updating homepage
When a user configures `confluence_root_homepage` to adjust the homepage
for a personal space on Confluence Cloud, the instance will implicitly
attempt to adjust the space type to `global` instead of its original
`personal` type. Confluence will automatically fail the REST attempt
with the message:

    Cannot change space type from personal to global or vice versa

To prevent this, explicitly configure the space's type field to the
original once detected when querying the space information during
initialization.

Signed-off-by: James Knight <james.d.knight@live.com>