remove flagged error state with retry-after header
Initial implementation [1] of rate limiting made an assumption that the
existence of the `Retry-After` header was an indication of an error. The
header can be defined in a success case (which is assumed it could be
used to help limit subsequent success cases). For now, remove the
incorrect check by only applying rate limiting on the 429 error.

A future change can better handle rate limiting before the 429 state is
triggered.

[1]: df2d6d482c6ff46cce145daff9bc03c690243230

Signed-off-by: James Knight <james.d.knight@live.com>