support rendering math using latex macros
Adds the ability for users to tailor their documentation generation to
use a Confluence-provided LaTeX macro. Confluence does not support LaTeX
by default; however, a Confluence instance may support LaTeX if a
marketplace add-on provides a macro supporting this.

This commit processing math blocks in a similar fashion as before --
math content is defined in a document, the content is converted into
LaTeX and this extension will use the imgmath extension to generate
images. This change adds some conditions to stop the conversion from
LaTex to images if a user specifies a `confluence_latex_macro` option.
When set, a macro will be generated instead (using the value of this
option as the name of the macro).

Signed-off-by: James Knight <james.d.knight@live.com>