adjust publisher tests to prevent connection reset states
The initial implementation of the publisher tests had a high chance in
timing out if the mocked HTTP server failed to process requests in a
timely fashion (e.g. connection error 54). The main issue was that the
HTTP server could only handle a single request at a time and an fixed
timeout of one second is enforced on these tests. This would lead to
cases where a unit test would issue some requests, the request were
being handled with a response pushed back, and a trailing request could
automatically kick in before the mocked instance formally closed the
request. If the unit testing (client) did not get the request served
right away in the first second, it would timeout.

Improvements have been added to the test -- we relax the timeout a bit
(to ensure the GitHub action instances have a bit more flexibility),
enable publishing logging to help better understand error cases if they
occur again, ensure request/responses are properly synchronized between
a unit test (client) and mocked server instance, and adjust the mocked
HTTP instance to handle requests in an aynchronous fashion.

Signed-off-by: James Knight <james.d.knight@live.com>