improve rate limiting handling
The following reworks the rate-limit handling this extension performs
when interacting with a Confluence instance. The initial implementation
would wait for a 429 error before attempting to limit calls, instead of
proactively delaying API calls which Confluence hints the client to
perform. The new implementation will now check for the retry-after
header option after each API request made, and apply this delay on a
subsequent API call.

This commit reworks the Rest and related implementation a bit. Instead
of using a `RestRateLimited` class to capture failed API requests, the
implementation has been switched to using decorators. This cleans up the
implementation a bit and helps manage rate limiting data all in the
`Rest` class.

In addition, a `requests_exception_wrappers` decorator was introduced as
well to minimize the common implementation for the various REST calls.

Signed-off-by: James Knight <james.d.knight@live.com>