rest: parse expected retry-after values
Rate-limiting changes [1] attempted to properly track the `Retry-After`
header value to apply to any request made, instead of relying on the
self-managed exponential back-off implementation. The issue with this
first implementation is that is did not properly handle the possible
value types of the value -- specifically, not properly handling the
string value to an integer seconds value, and also failing to handle a
possible string DateTime value [2].

This commit attempts to handle all possible `Retry-After` values and
properly prepare a seconds delay value to apply to a next request.

[1]: d663b0f69506f631b3c02a3fec0f1811e43e9872
[2]: https://datatracker.ietf.org/doc/html/rfc2616.html#section-14.37

Signed-off-by: James Knight <james.d.knight@live.com>