drop macro restriction support
Restricting subset of macros is an undocumented feature. It was added to
help deal with scenarios with a Confluence instance may have some
macros disabled by an administrator; providing a way for users to still
generate/publish documentation if a specific macro was disabled.

Newer Confluence instances do not support disabling core macros with
standard installs anymore ~ this includes all Confluence instances that
are supported by this extension (LTS versions). Since the likelihood
that these core macros are never disabled and that this feature was
undocumented, removing support for it. This should reduce some
implementation to minimize maintenance and validation time.

Signed-off-by: James Knight <james.d.knight@live.com>