translator: correct code-block's caption check
When tracking a caption value to apply to a code block, a check exists
to exclude processing this for a parsed literal (since this is not
supported). An issue with the initial implementation is that this
sanity check used the caption node to sanity check if a node was a
parsed literal block, instead of the sibling code block.

This commit corrects the implementation by checking for a parsed literal
block type on the sibling.

Signed-off-by: James Knight <james.d.knight@live.com>