builder: skip restrict ancestors check on dry-run without existing root
When configured for a dry-run state and a restrict ancestors check is
performed, if the root page did not already exist (to be updated), no
root page will exist to compiler a list of ancestors on. This causes
the extension to issue an invalid API request with a `None` page ID
call. Tweaking the restrict ancestors check to be skipped if no root
page is detected.

Signed-off-by: James Knight <james.d.knight@live.com>