support loading default local certification authority information
This extension primarily relies on the Requests module to manage
requests to a Confluence instance, which includes secure pipes. Requests
relies on embedded certificate information, which may not be ideal for
environments which have more updated certificates.

The following applies the internally managed SSL context to always be
registered; and if no configuration-provided client certificate is set,
certificates will be loaded from default locations for the running
platform (via `load_default_certs` [1]).

[1]: https://docs.python.org/3/library/ssl.html#ssl.SSLContext.load_default_certs

Signed-off-by: James Knight <james.d.knight@live.com>