track configuration options for configuration type conversions
A subset of options to be used as a boolean or integer type can be
accepted as a string (typically, from command line) and converted after
post-processing configuration options. The original implementation
defined a list of, for example, boolean configuration options and cycled
through these keys to ensured all string-assigned values are converted
to boolean types.

A maintenance issue observed is that as boolean options are defined for
the extension, development may miss adding these new options into the
respective type conversion lists in the "defaults" implementation. To
prevent this from happening, rework this extension's configuration
processing to help easily track the expected types for individual
configurations. With this change, the "defaults" implementation can
apply type conversions based off tracked configuration entries
registered in the manager.

Signed-off-by: James Knight <james.d.knight@live.com>