accept configuration options from environment
If a Confluence supported configuration option is not configured, allow
accepting an option from the running environment. Provides flexibility
for users wishing to configure a subset of options using environment
variables, instead of passing options via the command line or trying
to manipulate a project's `conf.py` file to read options from the
envionment.

Signed-off-by: James Knight <james.d.knight@live.com>