intersphinx: unrestrict domains
When intersphinx support was added, domains were restricted to only the
`std` domain. This extension only has performed limiting testing of its
intersphinx capabilities, and the limit was used to help prevent
possible issues from trying to processing unknown domain types. In
theory though, all domain types would like to be supported, and adding
a restriction makes it one extra step to deal with before trying to
correct/improve possible issues for intersphinx domains that may not
work. Removing the restriction to give users the capability of building
an inventory with all domain types. From there, if a user experiences
issues linking against a specific domain/reference, that example can be
later used as a test case to be addressed in a future version.

Signed-off-by: James Knight <james.d.knight@live.com>