default publish intersphinx databases
When an inventory can be created for documentation, this extension
leaves the file inside the output folder for a user to manage as they
see fit. Since these inventories are typically small, there is not a big
reason why we would need to prevent attaching them (by default) on a
root document. This commit changes the extension to support this.

The `confluence_publish_intersphinx` configuration option has been added
to allow a user to restrict publishing this file (by configuring it to
`False`).

In theory, a user will have a place to easily download the inventory
when building links between documentation. At this time, directly
mapping to an inventory file directly hosted on a Confluence instance
has two issues. First, authentication, which may not be easily
configured with intersphinx for some scenarios. Second, Confluence will
respond with a `Location` header to the static hosted file of the
inventory; however, this can cause the intersphinx to assume this is the
base URL for links (which causes badly formed links). Both of these
issues are primarily intersphinx limitations, which would need changes
in the intersphinx extension to improve support for these issues.

Signed-off-by: James Knight <james.d.knight@live.com>