use docutils node's findall call when available
As of docutils v0.18.1, a node now has a `findall` method (iterator)
over a now obsolete `traverse` method (list). This causes users using
the more recent docutils to have several deprecated warnings. To help
deal with users on newer and older versions, provide a helper call to
call `findall` if it is available.

Signed-off-by: James Knight <james.d.knight@live.com>