doc: remove optional hint in home description
We do not yet support building in one environment and publishing in
another environment, which may be inferred upon with the suggestion that
publishing is optional (disabling publishing is more so for checking
builds). To prevent any assumptions by a user, we will omit the use
of the optional term.

Signed-off-by: James Knight <james.d.knight@live.com>