drop pre-sphinx 4.3 support
Dropping a series of implementation used to support running with Sphinx
previous to v4.3. This extension aims to support the latest five
versions of Sphinx. An exception existed for supporting Sphinx v1.8 to
maintain support for a Python 2.7 interpreter. Since Python 2.7 is no
longer supported, dropping the older Sphinx support as well.

Signed-off-by: James Knight <james.d.knight@live.com>