publisher: remove archived page on re-introduction
If a publish event attempts to upload a page with a same name of an
archived page, the publish attempt will fail (due to conflict). There
currently is no way to unarchive a page through the REST API interface
at this time. To help deal with these scenarios, attempt to remove the
page and then upload the new page contents.

This will not work for users who do not have delete support; however,
users in this scenario would still fail to publish due to a page being
archived. At this time, the only workaround is to manually unarchive the
conflicting page(s).

Signed-off-by: James Knight <james.d.knight@live.com>