default enable hierarchy mode
The use of hierarchy mode is a more common configuration (utilizing
Sphinx's toctree definition to group/nest documents). It has been
discussed for some time that the hierarchy mode should be default
enabled -- and with the active development branch planning to a v2.0
release on a next release, the mode is being default enabled.

Signed-off-by: James Knight <james.d.knight@live.com>