improve paragraph indentation/offsets with v2 editor
An inconsistency has been reported when comparing indentations/offsets
between the original (v1) editor and new editor. First, when using a
v2 editor, line block types stacked against a paragraph would not have
the same spacing compared to the older editor. The use of the div tag
does not gracefully work in a v2 editor, and will be switched to a p
tag. This should help ensure expected spacing.

Next, for line blocks, the Confluence output should be indented these
configured blocks if detected that this are new line blocks inside, for
example, a paragraph. While this logic was applied on the v1 editor, the
same logic was not used on the v2 editor. We now apply a line-block
indentation if the condition is detected.

Signed-off-by: James Knight <james.d.knight@live.com>