restrict publishing with empty allow list
Users taking advantage of `confluence_publish_allowlist` may be using
this feature to detect which documents have changed to decide what
documents are to be published. In such systems, if an allow-list is
generated which is empty, the resulting state would be that all
documents will be published. This can be counter-intuitive for users
of this feature. This commit tweaks the allow list setup where users can
configure an empty allow list to indicate no pages will be published.

The default state, `confluence_publish_allowlist` is set to `None` to
not use this feature (i.e. all pages can be published). If an empty list
is assigned to this value:

    confluence_publish_allowlist = []

No documents will be published. This also applies to when using a file:

    confluence_publish_allowlist = 'allow-list.txt'

And if this file is empty, no files will be published.

Users may use the command line to override a persisted configuration by
setting `-Dconfluence_publish_allowlist=` (no value), to disable the
allow-list option. And if users need to support a command line override
hint to not publish any documents for their run state, users can still
override the `-Dconfluence_publish=0` option to disable publishing.

Signed-off-by: James Knight <james.d.knight@live.com>