support external referenced excerpts
The initial excerpt-include macros would rely on specifying document
names of pages which hold excerpts (to ensure an excerpt-include could
link to an excerpt). A user may want more freedom in the selection of an
excerpt, either by specifying the document title value themselves, or
even defining the "space:docname" format also supported by Confluence.

To support both worlds, excerpt-includes with a prefixed with `!` will
assume the reference is a local docname to reference. Otherwise, the
full value will be used to reference a page on the Confluence instance.

Signed-off-by: James Knight <james.d.knight@live.com>