builder: provide tempdir for ext-imgmath
For users building with Windows and using `dvisvgm` from MiKTeX, the
process may fail when dealing with temporary file locations that do
not share a common partition as the output directory.

 ERROR: Windows API error 87: The parameter is incorrect.

The imgmath extension allows a builder to override where temporary
files are build -- use this to hint to using a temporary directory
on the same partition the output directory to help prevent issues.

Signed-off-by: James Knight <james.d.knight@live.com>