update source headers
The following updates the headers for a series of implementation files.
The purpose is two parts -- first, this update removes the coding hint
on all Python files. This was required when this extension supported
Python 2, which support was dropped in the latest release. Second, this
commit updates the implementation to use SPDX license identifiers in
its implementation.

Signed-off-by: James Knight <james.d.knight@live.com>