suppress flake8-bandit for use of sha1
This implementation uses the sha1 algorithm to generate a small hash
(which is later truncated smaller) to help provide users with a higher
probability of unique/fixed title names for publication. This use of
this hash is non-security-related and with that, adding a suppression
if a developer is linting with the flake8-bandit checks.

Signed-off-by: James Knight <james.d.knight@live.com>