publisher: ensure new pages include a configured version comment
When a user configures `confluence_version_comment`, the value used will
only show in a page's history for page updates. This is due to the
original implementation only populate the version message for page
updates. This commit adjusts the implementation to always populate a
`version` entry for a page with a configured message value.

Note that new page comments only appear to work on Confluence Cloud.
Tests with Confluence server (v7.19.x) does not appear to accept a
message value (ignored?).

Signed-off-by: James Knight <james.d.knight@live.com>