translator: linkcode v2 editor support
For `sphinx.ext.linkcode` references, the original handling of these
nodes would right-align the links for the "[source]" references. Using
a div-float for this alignment is not supported with Confluence's new
editor style. For an initial rendering support for this preliminary
linkcode support, ensure the link is place appropriately on the same
link as a autodoc line description, and that the content is properly
spaced.

Signed-off-by: James Knight <james.d.knight@live.com>