builder: enable preliminary linkcode support
For `sphinx.ext.linkcode` extension, prospect changes [1] will provide
support for other Sphinx extensions to take advantage of
linkcode-generated references. Updating this extension to hint that it
will support `sphinx.ext.linkcode`. This should help ensure a future
release the Confluence Builder extension will automatically be ready to
utilize the Sphinx-managed extension (if Sphinx does update to include
linkcode support for custom builders).

If Sphinx changes the way of integrating with linkcode or opts to not
support such an extension, the introduced line in this commit can be
modified or dropped at a later time.

[1]: https://github.com/sphinx-doc/sphinx/pull/10585

Signed-off-by: James Knight <james.d.knight@live.com>