tests: automatically find test container output name
When generating an output folder for specific unit tests, instead of
relying on a specific frame stack offset, search for the first `test_`
prefixed call. This should avoid to need various internal offsets,
avoid the need to override the offset if a test using a non-trivial
build setup and simply the implementation.

Signed-off-by: James Knight <james.d.knight@live.com>