translator: support transform empty/none handling
While `confluence_lang_transform` can be used by users to handle or
override specific language types, a user may still want to rely on
"regular" language types to be processed. Tweaking the use of a
configured transform to fallback to the regular language mapping calls
if a transform results an empty or `None` value.

Signed-off-by: James Knight <james.d.knight@live.com>