configurable cleanup search modes
Provides the ability for a user to configure how to search for legacy
pages when a cleanup option is configured. When a project is configured
to cleanup old documents, this extension will search for any legacy
pages found on the Confluence instance. As it uploads new documents (and
attachments), it will update the internal list of tracked documents to
indicate which ones are active. At the end of the process, the remaining
set of documents/attachments are considered legacy and will be cleaned
up.

A problem observed is that sometimes Confluence does not always report
expected page dependencies through the API. Originally, this
implementation would query specific pages descendants; however, after
initial testing, not all descendants would always appear. The
implementation was switched to use a search API instead, which seemed to
provide more consistent results (~5 years ago). Unfortunately, a search
request did not always show expected results. A hidden "aggresive"
search option was added to force search for descendants on each detect
page, which always seemd to give us the expected page count, but this
took a lot of API calls.

Since it is hard to know the "right" the way of searching for legacy
pages, this commit now exposes the option for a user to force pick an
option which may best suit them. We default to "direct" searching, which
"should" be the preferred way of searching for descendants. And if users
have issues, they may chose to opt to using a search or aggresive
variant search style.

Signed-off-by: James Knight <james.d.knight@live.com>