publisher: retry attachment upload on 503 error
It has been observed on Confluence Cloud that in some cases when a user
publishes a new attachment to a page (and maybe specifically attached to
a newly created page), Confluence may report a 503 error. The behaviour
is odd since Confluence does partially process the new attachment (a
viewable entry on the page's list of attachments), but the data for the
attachment entry is corrupted. And in a next publish attempt, since the
comment hash is unchanged, this extension will not attempt to re-upload.
To help prevent this case, if a 503 error state is detected, check to
see if the attachment entry was created with corrupted data (i.e. can we
query for an existing attachment). If we find it, re-attempt to publish
the attachment.

Signed-off-by: James Knight <james.d.knight@live.com>