publisher: generate warning on restricted attachment types
For environments using the marketplace extension Akeles Consulting's
"Attachment Checker for Confluence", certain files will be not permitted
as attachments for a documentation's publish event. While common file
types may work, most of these environments will probably not be
configured to accept Sphinx's Intersphinx file. This will result in
publication always failing (unless `confluence_publish_intersphinx` is
configured).

In order to be flexible for such environments, we will take a relaxed
approach for restricted files. In the event that an attachment fails due
to this "Attachment Checker" addon, we will generate a warning instead
of failing.

Note that this is only a crude tweak to this extension. The check for
this error case relies on the default error message for the addon. For
environments that rename this error message, this tweak will not work,
but there is no other immediate way to distinguish this error case.

Signed-off-by: James Knight <james.d.knight@live.com>