config: change cleanup search mode back to default
When introducing the ability for users to configure the cleanup search
mode, the default mode was switch from `search` to `direct`, with the
assumption that `direct` being the best candidate based on the API's
description. However, it has been observed through release testing that
Confluence Server/DC versions (e.g. v8.2.1), the API may fail with an
"unimplemented" error.

To prevent users from observing this issue with a stock load, restoring
the default back to a `search` mode.

[1]: 2b6bc487218f6555c74eb1ad61659daeb48707d6

Signed-off-by: James Knight <james.d.knight@live.com>