translator: support anchor links in v2 editor
Instead of defining internal page links using a `ac:link` macro, it
appears that in Confluence's v2 editor, it is expected to use plain `a`
tags to point to these anchors. This is in part due to anchor macros
added to a page will use the raw anchor value given, instead of
prefixing the anchors with page prefixes.

Signed-off-by: James Knight <james.d.knight@live.com>